+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/preprocess/DirectivePreprocessor.java
@@ -185,6 +185,7 @@ public class DirectivePreprocessor extends AbstractMessageEmitter implements Ast
         map.put("goal",                 node -> setEnumOption(node, GenerationGoal::byName, profile::setGoal, GenerationGoal::allowedValues));
         map.put("instruction-limit",    node -> setIntOption(node, profile::setInstructionLimit, 1, CompilerProfile.MAX_INSTRUCTIONS));
         map.put("link-guards",          node -> setBooleanOption(node, profile::setLinkedBlockGuards));
+        map.put("mlog-indent",          node -> setIntOption(node, profile::setMlogIndent, 0, CompilerProfile.MAX_MLOG_INDENT));
         map.put("optimization",         node -> setEnumOption(node, OptimizationLevel::byName, profile::setAllOptimizationLevels, OptimizationLevel::allowedValues));
         map.put("passes",               node -> setIntOption(node, profile::setOptimizationPasses, 1, CompilerProfile.MAX_PASSES));
         map.put("print-unresolved",     node -> setEnumOption(node, FinalCodeOutput::byName, profile::setFinalCodeOutput, FinalCodeOutput::allowedValues));